chore(front): Move layout file to index.tsx
