+++ client/src/stories/layout/layout.stories.tsx
@@ -1,5 +1,5 @@
 import type { ComponentMeta, ComponentStoryObj } from "@storybook/react";
-import Layout from "~/components/layout/layout";
+import Layout from "~/components/layout";
 import React from "react";
 
 type T = typeof Layout;
